Add my Duck (#428)
* Add Duck image

🦆 Duck Image
This image showcases a cute, vibrant duck with a bright yellow body and a cheerful expression. The duck is standing with its tiny orange feet, and its small beady eyes add to its playful charm. This image represents a fun and lighthearted touch to the project, symbolizing curiosity and enthusiasm

* Update contributors.js

* Update contributors.js